pyproject.toml
[project]
name = "prophit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-fasthtml>=0.6.9",
    "duckdb>=0.9.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
    "sqlalchemy>=2.0.41",
]
